% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hmmibd_c}
\alias{hmmibd_c}
\title{hmmIBD}
\usage{
hmmibd_c(param_list)
}
\arguments{
\item{param_list}{A list of parameters created with \code{hmm_ibd}}
}
\description{
hmmIBD implementation from https://github.com/glipsnort/hmmIBD
}
\details{
\code{hmmibd_c} implements hidden Markov model for detecting segments
  of shared ancestry (identity by descent) in genetic sequence data.
}
