% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{hmmibdr}
\alias{hmmibdr}
\alias{hmmibdr-package}
\title{Wrapper for HMMIBD in Rcpp}
\description{
Wrapper for HMMIBD in Rcpp
}
\details{
Rcpp implementation of THMMIBD
}
\references{
https://github.com/glipsnort/hmmIBD
}
