% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm_ibd.R
\name{hmm_ibd}
\alias{hmm_ibd}
\title{The Real McCOIL categorical method function}
\usage{
hmm_ibd(input_file, output_file, allele_freqs = NULL,
  genotypes_sec_pop = NULL, allele_freqs_sec_pop = NULL,
  max_fit_iterations = NULL, exclude_ids = NULL, analysis_ids = NULL,
  num_gens = NULL, overwrite = FALSE, fract_only = FALSE,
  eps = 0.001, min_inform = 10, min_discord = 0, max_discord = 1,
  nchrom = 14, min_snp_sep = 5, rec_rate = 7.47e-05)
}
\arguments{
\item{input_file}{File of genotype data. See below for format.}

\item{output_file}{Output file name. Two output files will be produced, with
".hmm.txt" and ".hmm_fract.txt" appended to the supplied name.}

\item{allele_freqs}{File of allele frequencies for the sample population.
Format: tab-delimited, no header, one variant per row.
Line format: <chromosome (int)> <position (bp, int)> <allele 1 freq> <all 2 freq> [<all 3 freq>] ...
The genotype and frequency files must contain exactly the same variants,
in the same order. If no file is supplied, allele frequencies are
calculated from the input data file.}

\item{genotypes_sec_pop}{File of genotype data from a second population;
same format as for -i. (added in 2.0.0)}

\item{allele_freqs_sec_pop}{File of allele frequencies for the second
population; same format as for -f. (added in 2.0.0)}

\item{max_fit_iterations}{Maximum number of fit iterations (defaults to 5).}

\item{exclude_ids}{File of sample ids to exclude from all analysis.
Format: no header, one id (string) per row.
(Note: b stands for "bad samples".)}

\item{analysis_ids}{File of sample pairs to analyze; all others are not
processed by the HMM (but are still used to calculate allele frequencies).
Format: no header,    tab-delimited, two sample ids (strings) per row.
(Note: "g" stands for "good pairs".)}

\item{num_gens}{Cap on the number of generations (floating point).
Sets the maximum value for that parameter in the fit. This is useful
if you are interested in recent IBD and are working with a population
with substantial linkage disequilbrium. Specifying a small value will
force the program to assume little recombination and thus a low
transition rate; otherwise it will identify the small blocks of LD as
ancient IBD, and will force the number of generations to be large.}

\item{overwrite}{Boolean detailing if output files already exist, should
they be overwritten. Deafult = FALSE}

\item{fract_only}{Boolean detailing whether to rturn just the fract.
Default = FALSE}

\item{eps}{Numeric for error rate in genotype calls. Default = .001}

\item{min_inform}{Minimum number of informative sites in a pairwise. Default = 10}

\item{min_discord}{Minimum discordance in comparison. Default = 0. Set > 0 to skip identical pairs}

\item{max_discord}{Maximum discordance in comparison. Default = 1. Set < 1 to skip unrelated pairs}

\item{nchrom}{Number of chromosomes. Default = 14 for falciparum}

\item{min_snp_sep}{Minimum snp distance, i.e. skip next snp(s) if too close to last one. Default = 5 (bp)}

\item{rec_rate}{Recombination rate. Default = 7.4e-7. (7.4e-5 cM/bp or 13.5 kb/cM Miles et al, Genome Res 26:1288-1299 (2016))}
}
\value{
return list of summary data frames of hmmIBD output
}
\description{
This function triggers the c code for the categorical method
}
