% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hmmidm_cpp}
\alias{hmmidm_cpp}
\title{hmmidm}
\usage{
hmmidm_cpp(param_list)
}
\arguments{
\item{param_list}{A list of parameters created with \code{hmm_idm}}
}
\description{
hmmIBD implementation from https://github.com/glipsnort/hmmIBD
}
\details{
\code{hmmidm_cpp} implements hidden Markov model for detecting segments
  of shared ancestry (identity by descent) in genetic sequence data.
}
